+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - fix: package name resolving for types referenced via a package-reference ('package:some_package/some_entrypoint.dart)
 - fix: private top level methods where treated as part of the public API (resulting in all types used there treated as part of the public API as well)
 - fix: carry over entry points of type aliases to the aliased type (especially important for private types)
+- fix: naming of unary operators was wrong (unaryunary...)
 
 ## Version 0.22.0
 - fix: Fixes an issue if we have to deal with two types with the same name